+++ mdw-workflow/src/META-INF/mdw/application-cache.xml
@@ -38,9 +38,6 @@
   <Cache Name="AssetRefCache" Type="ReadOnly">
     <Property Name="ClassName">com.centurylink.mdw.cache.impl.AssetRefCache</Property>
   </Cache>
-  <Cache Name="AppTokenCache" Type="ReadOnly">
-    <Property Name="ClassName">com.centurylink.mdw.cache.impl.AppTokenCache</Property>
-  </Cache>
   <!-- AllowableTasks is empty for non-VCS Assets, which use the classic MDWHub mechanism -->
   <Cache Name="AllowableTaskActions" Type="ReadOnly">
     <Property Name="ClassName">com.centurylink.mdw.services.task.AllowableTaskActions</Property>
